requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.3.0
Django==2.0.8
django-extensions==2.1.0
django-modelcluster==4.2
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.2
Fabric3==1.14.post1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
paramiko==2.4.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
python-decouple==3.1
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.23
wagtail==2.2.1
wcwidth==0.1.7
webencodings==0.5.1
Willow==1.1
